Remove lib
