+++ changewithin/__init__.py
@@ -1,4 +1,3 @@
 from __future__ import absolute_import
 from changewithin.changewithin import ChangeWithin
 from changewithin.changewithin import ChangeHandler
-from changewithin.lib import get_state
