+++ changewithin/changewithin.py
@@ -1,6 +1,8 @@
 from __future__ import absolute_import
 import os
 import re
+import sys
+from tempfile import mkstemp
 
 from configobj import ConfigObj
 import osmium
@@ -9,7 +11,7 @@ import gettext
 from jinja2 import Environment
 from osconf import config_from_environment
 import osmapi
-from lib import get_osc
+
 from raven import Client
 
 # Env vars:
@@ -19,6 +21,48 @@ from raven import Client
 # EMAIL_RECIPIENTS
 # EMAIL_LANGUAGE
 # CONFIG
+
+
+def get_state():
+    """
+    Downloads the state from OSM replication system
+
+    :return: Actual state as a str
+    """
+
+    r = requests.get('http://planet.openstreetmap.org/replication/day/state.txt')
+
+
+def get_osc(stateurl=None):
+    """
+    Function to download the osc file
+
+    :param stateurl: str with the url of the osc
+    :return: None
+    """
+
+    if not stateurl:
+        state = get_state()
+
+        # zero-pad state so it can be safely split.
+        state = '000000000' + state
+        path = '{0}/{1}/{2}'.format(state[-9:-6], state[-6:-3], state[-3:])
+        stateurl = 'http://planet.openstreetmap.org/replication/day/{0}.osc.gz'.format(path)
+
+    sys.stderr.write('downloading {0}...\n'.format(stateurl))
+    # prepare a local file to store changes
+    handle, filename = mkstemp(prefix='change-', suffix='.osc.gz')
+    os.close(handle)
+
+    with open(filename, "w") as f:
+        resp = requests.get(stateurl)
+        f.write(resp.content)
+    sys.stderr.write('Done\n')
+    #sys.stderr.write('extracting {0}...\n'.format(filename))
+    #os.system('gunzip -f {0}'.format(filename))
+
+    # knock off the ".gz" suffix and return
+    return filename
 
 
 class ChangeHandler(osmium.SimpleHandler):
